Updated Go version and build command
This commit updates the Set up Go step in the pipeline to use Go version 1.19, which is a newer version of Go than the one you were previously using.

Additionally, this commit updates the Build project step to include the -o main option, which specifies the name of the output binary as main.